projects/project6_2/dsl_service/requirements.txt
fastapi
uvicorn
pydantic
langgraph
langchain-core
pyyaml
